Add extension to Source to return [Track]
- This is what is async.
- Now [Track] is what is passed into output.
