+++ Sources/iTunes/Source+Tracks.swift
@@ -1,0 +1,19 @@
+//
+//  Source+Tracks.swift
+//
+//
+//  Created by Greg Bolsinga on 12/7/23.
+//
+
+import Foundation
+
+extension Source {
+  public func gather() async throws -> [Track] {
+    switch self {
+    case .itunes:
+      return try Track.gatherAllTracks()
+    case .musickit:
+      return try await Track.gatherWithMusicKit()
+    }
+  }
+}
